+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Dankerino
 
-- Minor: Allow for adjusting rate limit settings.
 - Major: Add 7TV emotes from Chatterino7 (Mm2PL/Dankerino#79)
+- Minor: Move Dankerino settings to a separate tab (Mm2PL/Dankerino#84)
+- Minor: Allow for adjusting rate limit settings.
 
 ### Chatterino